pyproject.toml
[tool.poetry]
name = "kaspi_parse"
version = "0.1.0"
description = ""
authors = ["Leo Trubach <leotrubach@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"pdfminer.six" = "^20220506"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
